pyproject.toml
[project]
name = "gigachess"
version = "0.1.0"
description = "Best service for chess"
authors = [
    {name = "gigasterk",email = "andrewgigaster@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "crypto (>=1.4.1,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
